create keys.go and blocked.go to make app.go easier
